REMOVED LINKS PARAMETER FROM New-SchoolUserPhone
Blackbaud's team found that this endpoint was not
actually designed to create that link with another user.
They have updated the example and the endpoint documentation to reflect this
as well as added a new work item for the team to add a
new endpoint to handle the linking functionality.
When this endpoint is released it will be announced in the change logs.